fix: fix bed selection on patient details
