+++ api-services/mutations/tasks/bed_mutations.ts
@@ -49,9 +49,9 @@ export const useBedCreateMutation = () => {
       return { id: res.getId(), name: bed.name }
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.beds]).catch(console.error)
-      queryClient.refetchQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
-      queryClient.refetchQueries([QueryKeys.wards]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.beds]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.wards]).catch(console.error)
     },
   })
 }
@@ -76,8 +76,8 @@ export const useBedUpdateMutation = () => {
       return obj
     },
     onSuccess: () => {
-      queryClient.refetchQueries([APIServices.bed]).catch(console.error)
-      queryClient.refetchQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
+      queryClient.invalidateQueries([APIServices.bed]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
     },
   })
 }
@@ -100,9 +100,9 @@ export const useBedDeleteMutation = () => {
       return obj
     },
     onSuccess: () => {
-      queryClient.refetchQueries([APIServices.bed]).catch(console.error)
-      queryClient.refetchQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
-      queryClient.refetchQueries([QueryKeys.wards]).catch(console.error)
+      queryClient.invalidateQueries([APIServices.bed]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.wards]).catch(console.error)
     },
   })
 }
