+++ api-services/mutations/tasks/task_mutations.ts
@@ -125,8 +125,8 @@ export const useTaskCreateMutation = (callback: (task: TaskDTO) => void = noop, 
       return newTask
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.tasks, QueryKeys.patients]).catch(console.error)
-      queryClient.refetchQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.tasks, QueryKeys.patients]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
     }
   })
 }
@@ -149,8 +149,8 @@ export const useTaskUpdateMutation = (callback: () => void = noop) => {
       return updateTask.toObject()
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.tasks]).catch(console.error)
-      queryClient.refetchQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.tasks]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
     }
   })
 }
@@ -167,8 +167,8 @@ export const useTaskDeleteMutation = (callback: () => void = noop) => {
       return req.toObject()
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.tasks]).catch(console.error)
-      queryClient.refetchQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.tasks]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.rooms, roomOverviewsQueryKey]).catch(console.error)
     }
   })
 }
@@ -197,7 +197,7 @@ export const useSubTaskAddMutation = (taskId: string | undefined, callback: (sub
       return req.toObject()
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.tasks]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.tasks]).catch(console.error)
     }
   })
 }
@@ -224,7 +224,7 @@ export const useSubTaskUpdateMutation = (taskId?: string, callback: (subtask: Su
       return req.toObject()
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.tasks]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.tasks]).catch(console.error)
     }
   })
 }
@@ -242,7 +242,7 @@ export const useSubTaskDeleteMutation = (callback: () => void = noop) => {
       return req.toObject()
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.tasks]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.tasks]).catch(console.error)
     }
   })
 }
@@ -267,7 +267,7 @@ export const useAssignTaskMutation = (callback: () => void = noop) => {
       return req.toObject()
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.tasks]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.tasks]).catch(console.error)
     }
   })
 }
@@ -292,7 +292,7 @@ export const useUnassignTaskMutation = (callback: () => void = noop) => {
       return req.toObject()
     },
     onSuccess: () => {
-      queryClient.refetchQueries([QueryKeys.tasks]).catch(console.error)
+      queryClient.invalidateQueries([QueryKeys.tasks]).catch(console.error)
     }
   })
 }
