+++ tasks/components/modals/TaskDetailModal.tsx
@@ -324,8 +324,6 @@ export const TaskDetailModal = ({
     }
   }, [data, taskId])
 
-  console.log(task)
-
   const {
     data: personalTaskTemplatesData,
     isLoading: personalTaskTemplatesIsLoading,
